Allow translation message keys containing '.'
If a locale messages key contains a '.' character, it must be escaped with a
backslash. Because the backslash character must itself be escaped with a
backslash, the proper way to write a key the includes a '.' character is:

{
	"My\\.Key": "Value"
}

Some Excel formula names include a '.' character, so creating a translation
value using that formula name as a key requires support for '.' in keys.